BLUR:95 fixed.

+++ src/blur-core/src/main/java/org/apache/blur/manager/writer/IndexImporter.java
@@ -13,13 +13,16 @@ import java.util.concurrent.locks.ReadWriteLock;
 
 import org.apache.blur.log.Log;
 import org.apache.blur.log.LogFactory;
+import org.apache.blur.manager.BlurPartitioner;
 import org.apache.blur.server.ShardContext;
 import org.apache.blur.store.hdfs.HdfsDirectory;
 import org.apache.blur.utils.BlurConstants;
+import org.apache.blur.utils.BlurUtil;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.Text;
 import org.apache.lucene.index.AtomicReader;
 import org.apache.lucene.index.AtomicReaderContext;
 import org.apache.lucene.index.CompositeReaderContext;
@@ -92,15 +95,35 @@ public class IndexImporter extends TimerTask implements Closeable {
         for (HdfsDirectory directory : indexesToImport) {
           LOG.info("Starting import [{0}], commiting on [{1}/{2}]", directory, shard, table);
           indexWriter.commit();
-          applyDeletes(directory, indexWriter);
-          LOG.info("Add index [{0}] [{1}/{2}]", directory, shard, table);
-          indexWriter.addIndexes(directory);
-          LOG.info("Finishing import [{0}], commiting on [{1}/{2}]", directory, shard, table);
-          indexWriter.commit();
-          Path dirPath = directory.getPath();
-          LOG.info("Cleaning up old directory [{0}] for [{1}/{2}]", dirPath, shard, table);
-          fileSystem.delete(dirPath, true);
-          LOG.info("Import complete on [{0}/{1}]", shard, table);
+          boolean isSuccess = false;
+          boolean isRollbackDueToException = false;
+          try {
+            isSuccess = applyDeletes(directory, indexWriter, shard);
+          }catch(IOException e){
+            LOG.error("Some issue with deleting the old index", e);
+            isSuccess = false;
+            isRollbackDueToException = true;
+          }
+          if (isSuccess) {
+            LOG.info("Add index [{0}] [{1}/{2}]", directory, shard, table);
+            indexWriter.addIndexes(directory);
+            LOG.info("Finishing import [{0}], commiting on [{1}/{2}]", directory, shard, table);
+            indexWriter.commit();
+            Path dirPath = directory.getPath();
+            LOG.info("Cleaning up old directory [{0}] for [{1}/{2}]", dirPath, shard, table);
+            fileSystem.delete(dirPath, true);
+            LOG.info("Import complete on [{0}/{1}]", shard, table);
+          } else {
+            if(!isRollbackDueToException){
+              LOG.error("Index is corrupted, RowIds are found in wrong shard [{0}/{1}], cancelling index import for [{2}]", shard, table, directory);
+            }
+            LOG.info("Starting rollback on [{0}/{1}]", shard, table);
+            indexWriter.rollback();
+            LOG.info("Finished rollback on [{0}/{1}]", shard, table);
+            Path oldDirPath = directory.getPath();
+            String newDirectoryName = oldDirPath.getName().split("\\.")[0] + ".bad_rowids";
+            fileSystem.rename(oldDirPath, new Path(oldDirPath.getParent(), newDirectoryName));
+          }
         }
       } finally {
         _lock.writeLock().unlock();
@@ -119,12 +142,15 @@ public class IndexImporter extends TimerTask implements Closeable {
     return result;
   }
 
-  private void applyDeletes(Directory directory, IndexWriter indexWriter) throws IOException {
+  private boolean applyDeletes(Directory directory, IndexWriter indexWriter, String shard) throws IOException {
     DirectoryReader reader = DirectoryReader.open(directory);
     try {
       LOG.info("Applying deletes in reader [{0}]", reader);
       CompositeReaderContext compositeReaderContext = reader.getContext();
       List<AtomicReaderContext> leaves = compositeReaderContext.leaves();
+      BlurPartitioner blurPartitioner = new BlurPartitioner();
+      Text key = new Text();
+      int numberOfShards = _shardContext.getTableContext().getDescriptor().getShardCount();
       for (AtomicReaderContext context : leaves) {
         AtomicReader atomicReader = context.reader();
         Fields fields = atomicReader.fields();
@@ -132,6 +158,13 @@ public class IndexImporter extends TimerTask implements Closeable {
         TermsEnum termsEnum = terms.iterator(null);
         BytesRef ref = null;
         while ((ref = termsEnum.next()) != null) {
+          byte[] rowIdInBytes = ref.bytes;
+          key.set(rowIdInBytes, 0, rowIdInBytes.length);
+          int partition = blurPartitioner.getPartition(key, null, numberOfShards);
+          int shardId = BlurUtil.getShardIndex(shard);
+          if( shardId != partition){
+            return false;
+          }
           Term term = new Term(BlurConstants.ROW_ID, BytesRef.deepCopyOf(ref));
           indexWriter.deleteDocuments(term);
         }
@@ -139,5 +172,6 @@ public class IndexImporter extends TimerTask implements Closeable {
     } finally {
       reader.close();
     }
+    return true;
   }
 }
